Migrating to Flask
From this commit we have started to implement Flask in all new versions of the platform, to learn more visit the wiki, in this commit we have started to create files for the proper functioning of Flask. ALL THE FILES OF THE OLD VERSION ARE TEMPORARILY IN /old, since the development has not yet been completed.